fix: 빌드에러 api 로직 수정
