+++ public/sw.js
@@ -1,94 +1,1 @@
-/**
- * Copyright 2018 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// If the loader is already loaded, just stop.
-if (!self.define) {
-  let registry = {};
-
-  // Used for `eval` and `importScripts` where we can't get script URL by other means.
-  // In both cases, it's safe to use a global var because those functions are synchronous.
-  let nextDefineUri;
-
-  const singleRequire = (uri, parentUri) => {
-    uri = new URL(uri + ".js", parentUri).href;
-    return registry[uri] || (
-      
-        new Promise(resolve => {
-          if ("document" in self) {
-            const script = document.createElement("script");
-            script.src = uri;
-            script.onload = resolve;
-            document.head.appendChild(script);
-          } else {
-            nextDefineUri = uri;
-            importScripts(uri);
-            resolve();
-          }
-        })
-      
-      .then(() => {
-        let promise = registry[uri];
-        if (!promise) {
-          throw new Error(`Module ${uri} didn’t register its module`);
-        }
-        return promise;
-      })
-    );
-  };
-
-  self.define = (depsNames, factory) => {
-    const uri = nextDefineUri || ("document" in self ? document.currentScript.src : "") || location.href;
-    if (registry[uri]) {
-      // Module is already loading or loaded.
-      return;
-    }
-    let exports = {};
-    const require = depUri => singleRequire(depUri, uri);
-    const specialDeps = {
-      module: { uri },
-      exports,
-      require
-    };
-    registry[uri] = Promise.all(depsNames.map(
-      depName => specialDeps[depName] || require(depName)
-    )).then(deps => {
-      factory(...deps);
-      return exports;
-    });
-  };
-}
-define(['./workbox-7144475a'], (function (workbox) { 'use strict';
-
-  importScripts();
-  self.skipWaiting();
-  workbox.clientsClaim();
-  workbox.registerRoute("/", new workbox.NetworkFirst({
-    "cacheName": "start-url",
-    plugins: [{
-      cacheWillUpdate: async ({
-        response: e
-      }) => e && "opaqueredirect" === e.type ? new Response(e.body, {
-        status: 200,
-        statusText: "OK",
-        headers: e.headers
-      }) : e
-    }]
-  }), 'GET');
-  workbox.registerRoute(/.*/i, new workbox.NetworkOnly({
-    "cacheName": "dev",
-    plugins: []
-  }), 'GET');
-  self.__WB_DISABLE_DEV_LOGS = true;
-
-}));
-//# sourceMappingURL=sw.js.map
+if(!self.define){let e,n={};const a=(a,s)=>(a=new URL(a+".js",s).href,n[a]||new Promise((n=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=n,document.head.appendChild(e)}else e=a,importScripts(a),n()})).then((()=>{let e=n[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(s,c)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(n[i])return;let r={};const t=e=>a(e,i),p={module:{uri:i},exports:r,require:t};n[i]=Promise.all(s.map((e=>p[e]||t(e)))).then((e=>(c(...e),r)))}}define(["./workbox-c06b064f"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/231-8cbc9f07fd7974e2.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/2433-a3c2272ce6e32cc3.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/2778-6a1712d23e611409.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/2967-86aef47a05f0281b.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/3771-0413875146e32755.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/4510-c67bd309bafa4e6c.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/5190-88107600b5c8b2fc.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/6111-fc803a4b575d4b6c.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/6612-1dc1b3cadee020c9.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/7116-60045fc3c0b0381f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/7271-938511c36ff30d26.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/8173-275b4573b7cac1f5.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/8318a8be-eb73eb9ebc18b091.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/8866-0a1ff8780df5aa87.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/9201-00d1a2dc6cadbaed.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/9646-15f0506087ab33c8.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/998-3aa0451c39cf29b9.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/(common)/(auth)/signin/page-188c6be32a2c5032.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/(common)/(auth)/signup/page-b23c39f1a66f07b5.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/(common)/explore/page-242c7b5385549f45.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/(common)/ranking/page-c93d975e8982288b.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/(common)/selectrole/page-b78e3f88cc670408.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/_not-found/page-c51cc89ad3b69e1f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/layout-4d6d1f30c22a0af9.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/page-33e307e1e914a553.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/(home)/layout-d8e0e5f330f96f2a.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/(home)/page-c75988fc008fc07a.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chatroom/%5BroomId%5D/page-314d54e4d62d6237.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/@tabs/before/page-0e1759bc71721ef5.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/@tabs/closed/page-1688c600c74325e6.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/@tabs/confirm/page-d57b5b8e3320c471.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/@tabs/layout-1005a64d5e16a5bc.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/@tabs/page-730f77685efde133.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/@tabs/pending/page-edfb26c42a2d5695.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/chats/layout-467d3c17ce934e8f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/cash/account/page-a4f34b266a5faa5f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/cash/exchange/page-13da77149c009ad5.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/cash/history/page-743082af1de86b23.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/cash/page-35519847d134e73d.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/menu/page-4ee09f20222e147b.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/page-e5d46161f2e4715f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/information/page-246dfd1996fd0744.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/layout-2d3bddbeed77ebcf.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/page-d9d3991f4cc84243.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/preview/@tabs/review/page-5b812bb2008564bc.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/preview/layout-a7a250bb8bef0ae6.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/profile/page-acf2f83789e99f2c.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/styles/%5BpostId%5D/page-a7383ebef18359c7.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/mypage/styles/edit/page-333a6bb5a628edde.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/partner/reqcoordi/%5BrequestId%5D/page-7cd55a42408dcf25.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/(home)/layout-88fb2890889ad891.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/(home)/page-03f7faa0744b1ead.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chatroom/%5BroomId%5D/page-a15af4b03ead4eeb.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/@tabs/before/page-d2a262405ebd93bb.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/@tabs/closed/page-40d30290b995d7f7.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/@tabs/confirm/page-5d44be48ab9d500d.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/@tabs/layout-d23768aca1f25c90.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/@tabs/page-b7a5d350633fed56.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/@tabs/pending/page-92a24a73e2bfcf8f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/chats/layout-92ce7654ce975c59.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/confirm/page-bd3f716d2821603e.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/posts/%5BpostId%5D/page-1319494b810a4b4a.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/information/page-a6bb47a53c84bfef.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/layout-bed3c5b849318491.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/page-c573c6e66767c445.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/@tabs/review/page-4e7edc4a2e896c01.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/profile/layout-23aa6a199925785f.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/reqcoordi/layout-50132c1041d49e06.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/%5BpartnerId%5D/reqcoordi/page-b82588f9477d45e0.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/coordinator/page-34518aab286072e6.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/page-732bbc5a3f7f85a8.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/profile/editbodyinfo/page-564058bf341ad4ce.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/profile/editsize/page-5e7312f23607ae16.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/profile/editstyle/page-95abdc7b35e8c97b.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/profile/layout-144bea54e50fa296.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/profile/page-751c3a6da2477eae.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/reqlist/%5Brequestid%5D/editrequest/page-ac38e510877ab673.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/reqlist/%5Brequestid%5D/page-137b24e1ba00927d.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/reqlist/layout-67c8cbb5091bbe65.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/reqlist/newreq/page-52c410e6a614b8ba.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/mypage/reqlist/page-779c14bce2b44fe1.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/payment/cash/charge/page-748f1699fa2e144a.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/app/user/payment/confirm/page-1d02445d04b227d2.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/fd9d1056-1d75990e62d644bb.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/framework-8e0e0f4a6b83a956.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/main-app-2cc063d09c3076b8.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/main-e91f53119d3d27d6.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/pages/_app-f870474a17b7f2fd.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/pages/_error-c66a4e8afc46f17b.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-ac4a612fe44dc3d3.js",revision:"pnmAJwxn-ncZyVIkh5PaG"},{url:"/_next/static/css/7bf773dac7d32baf.css",revision:"7bf773dac7d32baf"},{url:"/_next/static/css/8c363533df73ec85.css",revision:"8c363533df73ec85"},{url:"/_next/static/media/ff840cfebfb63b0c-s.p.woff2",revision:"302ec55f5b4320354ec6b35a53dead87"},{url:"/_next/static/pnmAJwxn-ncZyVIkh5PaG/_buildManifest.js",revision:"3e2d62a10f4d6bf0b92e14aecf7836f4"},{url:"/_next/static/pnmAJwxn-ncZyVIkh5PaG/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/bank/032.svg",revision:"a717346ebf03863ca777c31e5a5cd47b"},{url:"/firebase-messaging-sw.js",revision:"e5062b1654de7390726fd795cd9a59d4"},{url:"/icon-192x192.png",revision:"0846a77b2550492e534f7a1a7012eee8"},{url:"/icon-256x256.png",revision:"bac2fc762e53b57f2ce55e628e5c5be9"},{url:"/icon-384x384.png",revision:"e9d31a292c162deb204ee96297648157"},{url:"/icon-512x512.png",revision:"85feb3a273be45d763ea33b13ef941c8"},{url:"/icons/add-circle.svg",revision:"cb32e1a7838bd25298b2db63441097d5"},{url:"/icons/add-icon.svg",revision:"6f9683700bb6cdcd2feddb69e3b74c90"},{url:"/icons/alert.svg",revision:"53b457ffe1c1916e435d1cb6b0a98736"},{url:"/icons/blog.svg",revision:"55481482768728825523f8ea0fdc931e"},{url:"/icons/chats.svg",revision:"4105bffa4387a8c629b2679e1fcc9083"},{url:"/icons/delete-circle-red.svg",revision:"681fb4934a04d4071552589ee26ad0c3"},{url:"/icons/explore.svg",revision:"469da60dab0d375274bf8cf87f0bf51b"},{url:"/icons/facebook.svg",revision:"bf6860ea5647ee8e43edf4dce776777c"},{url:"/icons/hamburger.svg",revision:"cbfcd2b4b0845ebb32afe9f3c13de158"},{url:"/icons/hanger.svg",revision:"5130b0e1763a1e0516389a58be85ba18"},{url:"/icons/home.svg",revision:"c89088f6434a65d452e060c083099159"},{url:"/icons/instagram.svg",revision:"8c1d1ca2d3c2f7715d2adbc829f2869a"},{url:"/icons/kakaopay.svg",revision:"ff7e7cb122f01358878e98e294d9d134"},{url:"/icons/link.svg",revision:"2ed8638625cde7ffcbcfc2e456e547c8"},{url:"/icons/list-arrow.svg",revision:"bf0c12c878ec1ee4c55911008ba4e2ed"},{url:"/icons/menu-line.svg",revision:"3b4b04bdc82133b9142f5c883144f7af"},{url:"/icons/money-send.svg",revision:"f66333b18bdfbcd547aa051d74414099"},{url:"/icons/more-vertical.svg",revision:"efc0580ee3265c28a768f1dcb763a57d"},{url:"/icons/mypage.svg",revision:"7f3fd0a5b4015fee5c6257a7c068f4aa"},{url:"/icons/notifications.svg",revision:"f5557e4b6c5a9f39f0b74d0715d00c45"},{url:"/icons/npay.svg",revision:"a9f434e046ff13db5bc42461928854fd"},{url:"/icons/pig.svg",revision:"551c7bc96aeb36a25ff438f0597dd710"},{url:"/icons/ranking.svg",revision:"dde98b9bf0a72b0d1dbad0783983384e"},{url:"/icons/reminder.svg",revision:"452c22ae5b4e99ffe92dc5bf61a94196"},{url:"/icons/shop.svg",revision:"201a70812f3aabe13ab5a18e5fbea1c1"},{url:"/icons/smile-coin.svg",revision:"2521054e2e3b2fbf73097cb30eb25108"},{url:"/icons/tiktok.svg",revision:"174968eace91fd753881942b5e6279ca"},{url:"/icons/tosspay.svg",revision:"e2d26fb2d372b8e779dddb426d5f525c"},{url:"/icons/youtube.svg",revision:"15bfed8f7002d6ca54d323858ec05f8c"},{url:"/images/chat-add-file.svg",revision:"aff975e993ad7c4fc3ebb63cb8774e43"},{url:"/images/chat-send-button.svg",revision:"a2c1cb5c63655da881358e5b35070630"},{url:"/images/home.svg",revision:"b36b13ea11bc0c657091ebf3e4995fe4"},{url:"/images/mfc-logo.png",revision:"1509ef459129bc4ce20097d00fcdba46"},{url:"/images/my_image.jpg",revision:"2cc6b660848720a1e3f0cba59cc63bbb"},{url:"/images/partner-post.jpeg",revision:"600a5b6bbd785e3865ce3109e51e8cb0"},{url:"/images/partner-signup.svg",revision:"c8686aa20857250ff3c077962437b605"},{url:"/images/pencil-icon.svg",revision:"0077cf4128c546f06f5646eb24a03847"},{url:"/images/pencil.svg",revision:"7ec95ef405e963f8debfbf6fb423cea6"},{url:"/images/profile.png",revision:"b2186ebae126a379baa02be675111dc4"},{url:"/images/reference_image.jpg",revision:"3d3970ae7539268f3695be6595dde26c"},{url:"/images/test-profile.jpg",revision:"faeecadf6a5cec73918943eabf140f00"},{url:"/images/test.svg",revision:"9ee51ab61065d6c47ce4a8803a3976b4"},{url:"/images/user-signup-man.svg",revision:"0725c84c4c32396f4ccdbf96359b82d4"},{url:"/images/user-signup-woman.svg",revision:"ff804f663fe148de973c951c42a6147d"},{url:"/initial_page.svg",revision:"df58b339bac3d2e4699a930bc4aeb761"},{url:"/manifest.json",revision:"49377dff528b9ba7c1a926a3f5242dd3"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/vercel.svg",revision:"61c6b19abff40ea7acd577be818f3976"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e,url:{pathname:n}})=>!(!e||n.startsWith("/api/auth/callback")||!n.startsWith("/api/"))),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:n},sameOrigin:a})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&a&&!n.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:n},sameOrigin:a})=>"1"===e.headers.get("RSC")&&a&&!n.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:{pathname:e},sameOrigin:n})=>n&&!e.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e})=>!e),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
